- Reverted back to injecting dependencies on ViewModels - This allows for the activities not having to know about details on how to instance them. - Injecting ViewModels with DI framework is not ideal, will find a better solution in the future.
