+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             android:name=".view.detail.DetailActivity"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".view.list.ListActivity"
+            android:name=".view.list.WallpaperListActivity"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
