+++ app/src/main/java/com/pmatuki/wowpapers/WowpaperApplication.kt
@@ -2,6 +2,7 @@ package com.pmatuki.wowpapers
 
 import android.app.Application
 import android.content.Context
+import com.pmatuki.wowpapers.di.AppScope
 import toothpick.config.Module
 import toothpick.ktp.KTP
 import toothpick.ktp.binding.bind
@@ -10,17 +11,7 @@ class WowpaperApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
-        KTP.openRootScope().openSubScope(APP_SCOPE).installModules(AppModule(this)).inject(this)
+        AppScope(this).scope.inject(this)
     }
 
-    class AppModule(application: Application) : Module() {
-        init {
-            val context = application.applicationContext
-            bind<Context>().toInstance(context)
-        }
-    }
-
-    companion object {
-        const val APP_SCOPE = "WowpaperApplicationScope"
-    }
 }
