+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.list.ListActivity">
+    tools:context=".view.list.WallpaperListActivity">
 
     <FrameLayout
         android:id="@+id/layout_view_loading"
